requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-leaflet==0.29.1
GDAL==3.4.1
idna==3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
django-bootstrap-v5==1.0.11
